EHN QCFC by subject groups (#35)
* EHN draft out what has to be done

* EHN calculate qc-fc by group

* FIX make sure age and gender are floats

- FIX path in scripts for beluga
- make sure age and gender are floats

note: when refactoring the code, make sure use the fmriprep data
fetcher; should probably also aim for a data fetcher for the connectome
data

* FIX narval -> beluga

* narval->beluga

* FIX multilevel header was not saving correctly

* EHN plot qcfc by groupgit add fmriprep_denoise/git add fmriprep_denoise/!

* Split the job submission script even further....

* isolate schaefer7network 800 due to memory requirement

* TEST dataloader and more logs.

* FIX type error...

* separate the calculation of the two metrics

* FIX different index in connectome and phenotype

make sure the phenotype data and flatten connectome only contains
the same subjects

+++ fmriprep_denoise/data/fmriprep.py
@@ -152,5 +152,5 @@ def _phenotype_movement(data):
     covar.loc[covar['gender'] == 'F', 'gender'] = 1
     covar.loc[covar['gender'] == 'M', 'gender'] = 0
     covar['gender'] = covar['gender'].astype('float')
-
+    covar['age'] = covar['age'].astype('float')
     return pd.concat((group_mean_fd, covar), axis=1)
